More refactor.
