+++ recirq/qcqmc/__init__.py
@@ -17,10 +17,9 @@ from typing import Optional
 from cirq.protocols.json_serialization import DEFAULT_RESOLVERS, ObjectFactory
 
 from .fermion_mode import FermionicMode
-from .hamiltonian import (HamiltonianData, HamiltonianFileParams,
-                          PyscfHamiltonianParams)
-from .trial_wf import (LayerSpec, PerfectPairingPlusTrialWavefunctionParams,
-                       TrialWavefunctionData)
+from .hamiltonian import HamiltonianData, HamiltonianFileParams, PyscfHamiltonianParams
+from .layer_spec import LayerSpec
+from .trial_wf import PerfectPairingPlusTrialWavefunctionParams, TrialWavefunctionData
 
 
 @lru_cache()
