+++ recirq/qcqmc/optimize_wf.py
@@ -14,7 +14,7 @@
 
 import copy
 import itertools
-from typing import Callable, Dict, Iterator, List, Mapping, Optional, Sequence, Tuple
+from typing import Dict, List, Mapping, Optional, Sequence, Tuple, Union
 
 import cirq
 import fqe
@@ -28,73 +28,14 @@ import scipy.sparse
 
 from recirq.qcqmc import (
     afqmc_circuits,
+    afqmc_generators,
+    converters,
     data,
     fermion_mode,
     hamiltonian,
-    qubit_maps,
+    layer_spec,
     trial_wf,
 )
-
-
-def _get_reorder_func(
-    *,
-    mode_qubit_map: Mapping[fermion_mode.FermionicMode, cirq.Qid],
-    ordered_qubits: Sequence[cirq.Qid],
-) -> Callable[[int, int], int]:
-    """This is a helper function that allows us to reorder fermionic modes.
-
-    Under the Jordan-Wigner transform, each fermionic mode is assigned to a
-    qubit. If we are provided an openfermion FermionOperator with the modes
-    assigned to qubits as described by mode_qubit_map this function gives us a
-    reorder_func that we can use to reorder the modes (with
-    openfermion.reorder(...)) so that they match the order of the qubits in
-    ordered_qubits. This is necessary to make a correspondence between
-    fermionic operators / wavefunctions and their qubit counterparts.
-
-    Args:
-        mode_qubit_map: A dict that shows how each FermionicMode is mapped to a qubit.
-        ordered_qubits: An ordered sequence of qubits.
-    """
-    qubits = list(mode_qubit_map.values())
-    assert len(qubits) == len(ordered_qubits)
-
-    # We sort the key: value pairs by the order of the values (qubits) in
-    # ordered_qubits.
-    sorted_mapping = list(mode_qubit_map.items())
-    sorted_mapping.sort(key=lambda x: ordered_qubits.index(x[1]))
-
-    remapping_map = {}
-    for i, (mode, _) in enumerate(sorted_mapping):
-        openfermion_index = 2 * mode.orb_ind + (0 if mode.spin == "a" else 1)
-        remapping_map[openfermion_index] = i
-
-    def remapper(index: int, _: int) -> int:
-        """A function that maps from the old index to the new one.
-
-        The _ argument is because it's expected by openfermion.reorder"""
-        return remapping_map[index]
-
-    return remapper
-
-
-def _get_pp_plus_gate_generators(
-    *, n_elec: int, heuristic_layers: Tuple[trial_wf.LayerSpec, ...], do_pp: bool = True
-) -> List[of.FermionOperator]:
-    heuristic_gate_generators = get_heuristic_gate_generators(n_elec, heuristic_layers)
-    if not do_pp:
-        return heuristic_gate_generators
-
-    n_pairs = n_elec // 2
-    pair_gate_generators = get_pair_hopping_gate_generators(n_pairs, n_elec)
-    return pair_gate_generators + heuristic_gate_generators
-
-
-def _get_ansatz_qubit_wf(
-    *, ansatz_circuit: cirq.Circuit, ordered_qubits: Sequence[cirq.Qid]
-):
-    return cirq.final_state_vector(
-        ansatz_circuit, qubit_order=list(ordered_qubits), dtype=np.complex128
-    )
 
 
 def get_and_check_energy(
@@ -106,7 +47,21 @@ def get_and_check_energy(
     one_body_basis_change_mat: np.ndarray,
     params: trial_wf.PerfectPairingPlusTrialWavefunctionParams,
 ) -> Tuple[float, float]:
-    ansatz_qubit_wf = _get_ansatz_qubit_wf(
+    """Compute the energy of the ansatz circuit and check against known values where possible.
+
+    Args:
+        hamiltonian_data: The Hamiltonian data.
+        ansatz_circuit: The ansatz circuit.
+        one_body_params: The one-body variational parameters.
+        two_body_params: The two-body variational parameters.
+        one_body_basis_change_mat: The one-body basis change matrix.
+        params: The trial wavefunction parameters.
+
+    Returns:
+        ansatz_energy: The total energy of the ansatz circuit.
+        hf_energy: The hartree-fock energy (initial energy of the ansatz circuit).
+    """
+    ansatz_qubit_wf = converters.get_ansatz_qubit_wf(
         ansatz_circuit=ansatz_circuit,
         ordered_qubits=params.qubits_jordan_wigner_ordered,
     )
@@ -127,7 +82,7 @@ def get_and_check_energy(
         one_body_params=one_body_params,
         two_body_params=two_body_params,
         wf=initial_wf,
-        gate_generators=_get_pp_plus_gate_generators(
+        gate_generators=afqmc_generators.get_pp_plus_gate_generators(
             n_elec=params.n_elec,
             heuristic_layers=params.heuristic_layers,
             do_pp=params.do_pp,
@@ -157,7 +112,16 @@ def build_pp_plus_trial_wavefunction(
     dependencies: Dict[data.Params, data.Data],
     do_print: bool = False,
 ) -> trial_wf.TrialWavefunctionData:
-    """Builds a TrialWavefunctionData from a TrialWavefunctionParams"""
+    """Builds a TrialWavefunctionData from a TrialWavefunctionParams
+
+    Args:
+        params: The parameters specifying the PP+ trial wavefunction.
+        dependencies: Data dependencies
+        do_print: Print debugging information to stdout
+
+    Returns:
+        The constructed TrialWavefunctionData object.
+    """
 
     if do_print:
         print("Building Trial Wavefunction")
@@ -165,9 +129,8 @@ def build_pp_plus_trial_wavefunction(
     hamiltonian_data = dependencies[params.hamiltonian_params]
     assert isinstance(hamiltonian_data, hamiltonian.HamiltonianData)
 
-    assert (
-        params.n_orb == params.n_elec
-    )  ## Necessary for perfect pairing wavefunction to make sense.
+    if params.n_orb != params.n_elec:
+        raise ValueError("PP wavefunction must have n_orb = n_elec")
 
     if params.do_optimization:
         (
@@ -195,11 +158,11 @@ def build_pp_plus_trial_wavefunction(
         n_one_body_params = params.n_orb * (params.n_orb - 1)
         one_body_params = np.zeros(n_one_body_params)
         one_body_basis_change_mat = np.diag(np.ones(params.n_orb * 2))
-        two_body_params = get_two_body_params_from_qchem_amplitudes(
+        two_body_params = converters.get_two_body_params_from_qchem_amplitudes(
             params.initial_two_body_qchem_amplitudes
         )
 
-    (superposition_circuit, ansatz_circuit) = get_circuits(
+    (superposition_circuit, ansatz_circuit) = afqmc_circuits.get_circuits(
         two_body_params=two_body_params,
         n_orb=params.n_orb,
         n_elec=params.n_elec,
@@ -237,6 +200,12 @@ def get_rotated_hamiltonians(
 ) -> Tuple[fqe_hams.RestrictedHamiltonian, float, scipy.sparse.csc_matrix]:
     """A helper method that gets the hamiltonians in the basis of the trial_wf.
 
+    Args:
+        hamiltonian_data: The specification of the Hamiltonian.
+        one_body_basis_change_mat: The basis change matrix for the one body hamiltonian.
+        mode_qubit_map: A mapping between fermionic modes and cirq qubits.
+        ordered_qubits: An ordered set of qubits.
+
     Returns:
         The hamiltonian in FQE form, minus a constant energy shift.
         The constant part of the Hamiltonian missing from the FQE Hamiltonian.
@@ -251,7 +220,7 @@ def get_rotated_hamiltonians(
     mol_ham.rotate_basis(one_body_basis_change_mat)
     fermion_operator_ham = of.get_fermion_operator(mol_ham)
 
-    reorder_func = _get_reorder_func(
+    reorder_func = converters.get_reorder_func(
         mode_qubit_map=mode_qubit_map, ordered_qubits=ordered_qubits
     )
     fermion_operator_ham_qubit_ordered = of.reorder(
@@ -274,9 +243,23 @@ def get_energy_and_check_sanity(
     mode_qubit_map: Mapping[fermion_mode.FermionicMode, cirq.Qid],
     ordered_qubits: Sequence[cirq.Qid],
 ) -> float:
-    """A method that checks for consistency and returns the ansatz energy."""
+    """A method that checks for consistency and returns the ansatz energy.
 
-    unrotated_fqe_wf_as_cirq = convert_fqe_wf_to_cirq(
+    Args:
+        circuit_wf: The cirq statevector constructed from the wavefunction ansatz circuit.
+        fqe_wf: The FQE wavefunction used for optimization.
+        unrotated_fqe_wf: The (unoptimized) unrotated FQE wavefunction.
+        fqe_ham: The restricted FQE hamiltonian.
+        sparse_ham: The qubit Hamiltonian as a sparse matrix.
+        e_core: The core energy (nuclear-repulision + any frozen core energy.)
+        mode_qubit_map: A mapping between fermionic modes and cirq qubits.
+        ordered_qubits: An ordered set of qubits.
+
+    Returns:
+        The ansatz energy.
+    """
+
+    unrotated_fqe_wf_as_cirq = converters.convert_fqe_wf_to_cirq(
         fqe_wf=unrotated_fqe_wf,
         mode_qubit_map=mode_qubit_map,
         ordered_qubits=ordered_qubits,
@@ -287,405 +270,11 @@ def get_energy_and_check_sanity(
     assert isinstance(ansatz_energy, float)
 
     fqe_energy = np.real(fqe_wf.expectationValue(fqe_ham) + e_core)
-    print(scipy.sparse.csc_matrix(circuit_wf))
-    print(scipy.sparse.csc_matrix(unrotated_fqe_wf_as_cirq))
     np.testing.assert_array_almost_equal(ansatz_energy, fqe_energy)
     np.testing.assert_array_almost_equal(
         circuit_wf, unrotated_fqe_wf_as_cirq, decimal=5
     )
     return ansatz_energy
-
-
-def get_4_qubit_pp_circuits(
-    *,
-    two_body_params: np.ndarray,
-    n_elec: int,
-    heuristic_layers: Tuple[trial_wf.LayerSpec, ...],
-) -> Tuple[cirq.Circuit, cirq.Circuit]:
-    """A helper function that builds the circuits for the four qubit ansatz.
-
-    We map the fermionic orbitals to grid qubits like so:
-    3 1
-    2 0
-    """
-    assert n_elec == 2
-
-    fermion_index_to_qubit_map = qubit_maps.get_4_qubit_fermion_qubit_map()
-    geminal_gate = afqmc_circuits.GeminalStatePreparationGate(
-        two_body_params[0], inline_control=True
-    )
-
-    ansatz_circuit = cirq.Circuit(
-        cirq.decompose(
-            geminal_gate.on(
-                fermion_index_to_qubit_map[0],
-                fermion_index_to_qubit_map[1],
-                fermion_index_to_qubit_map[2],
-                fermion_index_to_qubit_map[3],
-            )
-        )
-    )
-
-    heuristic_layer_circuit = get_heuristic_circuit(
-        heuristic_layers, n_elec, two_body_params[1:], fermion_index_to_qubit_map
-    )
-
-    ansatz_circuit += heuristic_layer_circuit
-
-    indicator = fermion_index_to_qubit_map[2]
-    superposition_circuit = cirq.Circuit([cirq.H(indicator) + ansatz_circuit])
-    ansatz_circuit = cirq.Circuit([cirq.X(indicator) + ansatz_circuit])
-
-    return superposition_circuit, ansatz_circuit
-
-
-def get_8_qubit_circuits(
-    *,
-    two_body_params: np.ndarray,
-    n_elec: int,
-    heuristic_layers: Tuple[trial_wf.LayerSpec, ...],
-) -> Tuple[cirq.Circuit, cirq.Circuit]:
-    """A helper function that builds the circuits for the four qubit ansatz.
-
-    We map the fermionic orbitals to grid qubits like so:
-    3 5 1 7
-    2 4 0 6
-    """
-    fermion_index_to_qubit_map = qubit_maps.get_8_qubit_fermion_qubit_map()
-
-    geminal_gate_1 = afqmc_circuits.GeminalStatePreparationGate(
-        two_body_params[0], inline_control=True
-    )
-    geminal_gate_2 = afqmc_circuits.GeminalStatePreparationGate(
-        two_body_params[1], inline_control=True
-    )
-
-    # We'll add the initial bit flips later.
-    ansatz_circuit = cirq.Circuit(
-        cirq.decompose(
-            geminal_gate_1.on(
-                fermion_index_to_qubit_map[2],
-                fermion_index_to_qubit_map[3],
-                fermion_index_to_qubit_map[4],
-                fermion_index_to_qubit_map[5],
-            )
-        ),
-        cirq.decompose(
-            geminal_gate_2.on(
-                fermion_index_to_qubit_map[0],
-                fermion_index_to_qubit_map[1],
-                fermion_index_to_qubit_map[6],
-                fermion_index_to_qubit_map[7],
-            )
-        ),
-    )
-
-    heuristic_layer_circuit = get_heuristic_circuit(
-        heuristic_layers, n_elec, two_body_params[2:], fermion_index_to_qubit_map
-    )
-
-    ansatz_circuit += heuristic_layer_circuit
-
-    superposition_circuit = (
-        cirq.Circuit(
-            [
-                cirq.H(fermion_index_to_qubit_map[0]),
-                cirq.CNOT(fermion_index_to_qubit_map[0], fermion_index_to_qubit_map[6]),
-                cirq.SWAP(fermion_index_to_qubit_map[0], fermion_index_to_qubit_map[4]),
-            ]
-        )
-        + ansatz_circuit
-    )
-
-    ansatz_circuit = (
-        cirq.Circuit(
-            [
-                cirq.X(fermion_index_to_qubit_map[4]),
-                cirq.X(fermion_index_to_qubit_map[6]),
-            ]
-        )
-        + ansatz_circuit
-    )
-
-    return superposition_circuit, ansatz_circuit
-
-
-def get_12_qubit_circuits(
-    *,
-    two_body_params: np.ndarray,
-    n_elec: int,
-    heuristic_layers: Tuple[trial_wf.LayerSpec, ...],
-) -> Tuple[cirq.Circuit, cirq.Circuit]:
-    """A helper function that builds the circuits for the four qubit ansatz.
-
-    We map the fermionic orbitals to grid qubits like so:
-    5 7 3 9 1 11
-    4 6 2 8 0 10
-    """
-
-    fermion_index_to_qubit_map = qubit_maps.get_12_qubit_fermion_qubit_map()
-
-    geminal_gate_1 = afqmc_circuits.GeminalStatePreparationGate(
-        two_body_params[0], inline_control=True
-    )
-    geminal_gate_2 = afqmc_circuits.GeminalStatePreparationGate(
-        two_body_params[1], inline_control=True
-    )
-    geminal_gate_3 = afqmc_circuits.GeminalStatePreparationGate(
-        two_body_params[2], inline_control=True
-    )
-
-    # We'll add the initial bit flips later.
-    ansatz_circuit = cirq.Circuit(
-        cirq.decompose(
-            geminal_gate_1.on(
-                fermion_index_to_qubit_map[4],
-                fermion_index_to_qubit_map[5],
-                fermion_index_to_qubit_map[6],
-                fermion_index_to_qubit_map[7],
-            )
-        ),
-        cirq.decompose(
-            geminal_gate_2.on(
-                fermion_index_to_qubit_map[2],
-                fermion_index_to_qubit_map[3],
-                fermion_index_to_qubit_map[8],
-                fermion_index_to_qubit_map[9],
-            )
-        ),
-        cirq.decompose(
-            geminal_gate_3.on(
-                fermion_index_to_qubit_map[0],
-                fermion_index_to_qubit_map[1],
-                fermion_index_to_qubit_map[10],
-                fermion_index_to_qubit_map[11],
-            )
-        ),
-    )
-
-    heuristic_layer_circuit = get_heuristic_circuit(
-        heuristic_layers, n_elec, two_body_params[3:], fermion_index_to_qubit_map
-    )
-
-    ansatz_circuit += heuristic_layer_circuit
-
-    superposition_circuit = (
-        cirq.Circuit(
-            [
-                cirq.H(fermion_index_to_qubit_map[8]),
-                cirq.CNOT(fermion_index_to_qubit_map[8], fermion_index_to_qubit_map[0]),
-                cirq.CNOT(fermion_index_to_qubit_map[8], fermion_index_to_qubit_map[2]),
-                cirq.SWAP(
-                    fermion_index_to_qubit_map[0], fermion_index_to_qubit_map[10]
-                ),
-                cirq.SWAP(fermion_index_to_qubit_map[2], fermion_index_to_qubit_map[6]),
-            ]
-        )
-        + ansatz_circuit
-    )
-
-    ansatz_circuit = (
-        cirq.Circuit(
-            [
-                cirq.X(fermion_index_to_qubit_map[6]),
-                cirq.X(fermion_index_to_qubit_map[8]),
-                cirq.X(fermion_index_to_qubit_map[10]),
-            ]
-        )
-        + ansatz_circuit
-    )
-
-    return superposition_circuit, ansatz_circuit
-
-
-def get_16_qubit_circuits(
-    *,
-    two_body_params: np.ndarray,
-    n_elec: int,
-    heuristic_layers: Tuple[trial_wf.LayerSpec, ...],
-) -> Tuple[cirq.Circuit, cirq.Circuit]:
-    """A helper function that builds the circuits for the four qubit ansatz.
-
-    We map the fermionic orbitals to grid qubits like so:
-    7 9 5 11 3 13 1 15
-    6 8 4 10 2 12 0 14
-    """
-    fermion_index_to_qubit_map = qubit_maps.get_16_qubit_fermion_qubit_map()
-
-    geminal_gate_1 = afqmc_circuits.GeminalStatePreparationGate(
-        two_body_params[0], inline_control=True
-    )
-    geminal_gate_2 = afqmc_circuits.GeminalStatePreparationGate(
-        two_body_params[1], inline_control=True
-    )
-    geminal_gate_3 = afqmc_circuits.GeminalStatePreparationGate(
-        two_body_params[2], inline_control=True
-    )
-    geminal_gate_4 = afqmc_circuits.GeminalStatePreparationGate(
-        two_body_params[3], inline_control=True
-    )
-
-    # We'll add the initial bit flips later.
-    ansatz_circuit = cirq.Circuit(
-        cirq.decompose(
-            geminal_gate_1.on(
-                fermion_index_to_qubit_map[6],
-                fermion_index_to_qubit_map[7],
-                fermion_index_to_qubit_map[8],
-                fermion_index_to_qubit_map[9],
-            )
-        ),
-        cirq.decompose(
-            geminal_gate_2.on(
-                fermion_index_to_qubit_map[4],
-                fermion_index_to_qubit_map[5],
-                fermion_index_to_qubit_map[10],
-                fermion_index_to_qubit_map[11],
-            )
-        ),
-        cirq.decompose(
-            geminal_gate_3.on(
-                fermion_index_to_qubit_map[2],
-                fermion_index_to_qubit_map[3],
-                fermion_index_to_qubit_map[12],
-                fermion_index_to_qubit_map[13],
-            )
-        ),
-        cirq.decompose(
-            geminal_gate_4.on(
-                fermion_index_to_qubit_map[0],
-                fermion_index_to_qubit_map[1],
-                fermion_index_to_qubit_map[14],
-                fermion_index_to_qubit_map[15],
-            )
-        ),
-    )
-
-    heuristic_layer_circuit = get_heuristic_circuit(
-        heuristic_layers, n_elec, two_body_params[4:], fermion_index_to_qubit_map
-    )
-
-    ansatz_circuit += heuristic_layer_circuit
-
-    superposition_circuit = (
-        cirq.Circuit(
-            [
-                cirq.H(fermion_index_to_qubit_map[10]),
-                cirq.CNOT(
-                    fermion_index_to_qubit_map[10], fermion_index_to_qubit_map[2]
-                ),
-                cirq.SWAP(
-                    fermion_index_to_qubit_map[2], fermion_index_to_qubit_map[12]
-                ),
-                cirq.CNOT(
-                    fermion_index_to_qubit_map[10], fermion_index_to_qubit_map[4]
-                ),
-                cirq.CNOT(
-                    fermion_index_to_qubit_map[12], fermion_index_to_qubit_map[0]
-                ),
-                cirq.SWAP(fermion_index_to_qubit_map[4], fermion_index_to_qubit_map[8]),
-                cirq.SWAP(
-                    fermion_index_to_qubit_map[0], fermion_index_to_qubit_map[14]
-                ),
-            ]
-        )
-        + ansatz_circuit
-    )
-
-    ansatz_circuit = (
-        cirq.Circuit(
-            [
-                cirq.X(fermion_index_to_qubit_map[8]),
-                cirq.X(fermion_index_to_qubit_map[10]),
-                cirq.X(fermion_index_to_qubit_map[12]),
-                cirq.X(fermion_index_to_qubit_map[14]),
-            ]
-        )
-        + ansatz_circuit
-    )
-
-    return superposition_circuit, ansatz_circuit
-
-
-def get_circuits(
-    *,
-    two_body_params: np.ndarray,
-    # from wf_params:
-    n_orb: int,
-    n_elec: int,
-    heuristic_layers: Tuple[trial_wf.LayerSpec, ...],
-) -> Tuple[cirq.Circuit, cirq.Circuit]:
-    """A function that runs a specialized method to get the ansatz circuits."""
-
-    # TODO(?): Just input one of these quantities.
-    if n_orb != n_elec:
-        raise ValueError("n_orb must equal n_elec.")
-
-    circ_funcs = {
-        2: get_4_qubit_pp_circuits,
-        4: get_8_qubit_circuits,
-        6: get_12_qubit_circuits,
-        8: get_16_qubit_circuits,
-    }
-    try:
-        circ_func = circ_funcs[n_orb]
-    except KeyError:
-        raise NotImplementedError(f"No circuits for n_orb = {n_orb}")
-
-    return circ_func(
-        two_body_params=two_body_params,
-        n_elec=n_elec,
-        heuristic_layers=heuristic_layers,
-    )
-
-
-def get_two_body_params_from_qchem_amplitudes(
-    qchem_amplitudes: np.ndarray,
-) -> np.ndarray:
-    """Translates perfect pairing amplitudes from qchem to rotation angles.
-
-    qchem style: 1 |1100> + t_i |0011>
-    our style: cos(\theta_i) |1100> + sin(\theta_i) |0011>
-    """
-
-    two_body_params = np.arccos(1 / np.sqrt(1 + qchem_amplitudes**2)) * np.sign(
-        qchem_amplitudes
-    )
-
-    # Numpy casts the array improperly to a float when we only have one parameter.
-    two_body_params = np.atleast_1d(two_body_params)
-
-    return two_body_params
-
-
-#################### Here be dragons.###########################################
-
-
-def convert_fqe_wf_to_cirq(
-    fqe_wf: fqe_wfn.Wavefunction,
-    mode_qubit_map: Mapping[fermion_mode.FermionicMode, cirq.Qid],
-    ordered_qubits: Sequence[cirq.Qid],
-) -> np.ndarray:
-    """Converts an FQE wavefunction to one on qubits with a particular ordering.
-
-    Args:
-        fqe_wf: The FQE wavefunction.
-        mode_qubit_map: A mapping from fermion modes to cirq qubits.
-        ordered_qubits:
-    """
-    n_qubits = len(mode_qubit_map)
-    fermion_op = fqe.openfermion_utils.fqe_to_fermion_operator(fqe_wf)
-
-    reorder_func = _get_reorder_func(
-        mode_qubit_map=mode_qubit_map, ordered_qubits=ordered_qubits
-    )
-    fermion_op = of.reorder(fermion_op, reorder_func, num_modes=n_qubits)
-
-    qubit_op = of.jordan_wigner(fermion_op)
-
-    return fqe.qubit_wavefunction_from_vacuum(
-        qubit_op, list(cirq.LineQubit.range(n_qubits))
-    )
 
 
 def get_one_body_cluster_coef(
@@ -759,282 +348,6 @@ def get_evolved_wf(
         )
 
     return rotated_wf, wf
-
-
-def get_pair_hopping_gate_generators(
-    n_pairs: int, n_elec: int
-) -> List[of.FermionOperator]:
-    """Get the generators of the pair-hopping unitaries.
-
-    Args:
-        n_pairs: The number of pair coupling terms.
-        n_elec: The total number of electrons.
-
-    Returns:
-        A list of gate generators
-    """
-    gate_generators = []
-    for pair in range(n_pairs):
-        to_a = n_elec + 2 * pair
-        to_b = n_elec + 2 * pair + 1
-        from_a = n_elec - 2 * pair - 2
-        from_b = n_elec - 2 * pair - 1
-
-        fop_string = f"{to_b} {to_a} {from_b}^ {from_a}^"
-
-        gate_generator = of.FermionOperator(fop_string, 1.0)
-        gate_generator = 1j * (gate_generator - of.hermitian_conjugated(gate_generator))
-        gate_generators.append(gate_generator)
-
-    return gate_generators
-
-
-def get_indices_heuristic_layer_in_pair(n_elec: int) -> Iterator[Tuple[int, int]]:
-    """Get the indicies for the heuristic layers.
-
-    Args:
-        n_elec: The number of electrons
-
-    Returns:
-        An iterator of the indices
-    """
-    n_pairs = n_elec // 2
-
-    for pair in range(n_pairs):
-        to_a = n_elec + 2 * pair
-        to_b = n_elec + 2 * pair + 1
-        from_a = n_elec - 2 * pair - 2
-        from_b = n_elec - 2 * pair - 1
-        yield (from_a, to_a)
-        yield (from_b, to_b)
-
-
-def get_indices_heuristic_layer_cross_pair(n_elec) -> Iterator[Tuple[int, int]]:
-    """Indices that couple adjacent pairs.
-
-    Args:
-        n_elec: The number of electrons
-
-    Returns:
-        An iterator of the indices
-    """
-    n_pairs = n_elec // 2
-
-    for pair in range(n_pairs - 1):
-        to_a = n_elec + 2 * pair
-        to_b = n_elec + 2 * pair + 1
-        from_next_a = n_elec - 2 * (pair + 1) - 2
-        from_next_b = n_elec - 2 * (pair + 1) - 1
-        yield (to_a, from_next_a)
-        yield (to_b, from_next_b)
-
-
-def get_indices_heuristic_layer_cross_spin(n_elec) -> Iterator[Tuple[int, int]]:
-    """Get indices that couple the two spin sectors.
-
-    Args:
-        n_elec: The number of electrons
-
-    Returns:
-        An iterator of the indices that couple spin sectors.
-    """
-    n_pairs = n_elec // 2
-
-    for pair in range(n_pairs):
-        to_a = n_elec + 2 * pair
-        to_b = n_elec + 2 * pair + 1
-        from_a = n_elec - 2 * pair - 2
-        from_b = n_elec - 2 * pair - 1
-        yield (to_a, to_b)
-        yield (from_a, from_b)
-
-
-def get_charge_charge_generator(indices: Tuple[int, int]) -> of.FermionOperator:
-    """Returns the generator for density evolution between the indices
-
-    Args:
-        indices: The indices to for charge-charge terms.:w
-
-    Returns:
-        The generator for density evolution for this pair of electrons.
-    """
-
-    fop_string = "{:d}^ {:d} {:d}^ {:d}".format(
-        indices[0], indices[0], indices[1], indices[1]
-    )
-    gate_generator = of.FermionOperator(fop_string, 1.0)
-
-    return gate_generator
-
-
-def get_charge_charge_gate(
-    qubits: Tuple[cirq.Qid, ...], param: float
-) -> cirq.Operation:
-    """Get the cirq charge-charge gate.
-
-    Args:
-        qubits: Two qubits you want to apply the gate to.
-        param: The parameter for the charge-charge interaction.
-
-    Returns:
-        The charge-charge gate.
-    """
-    return cirq.CZ(qubits[0], qubits[1]) ** (-param / np.pi)
-
-
-def get_givens_generator(indices: Tuple[int, int]) -> of.FermionOperator:
-    """Returns the generator for givens rotation between two orbitals.
-
-    Args:
-        indices: The two indices for the givens rotation.
-
-    Returns:
-        The givens generator for evolution for this pair of electrons.
-    """
-
-    fop_string = "{:d}^ {:d}".format(indices[0], indices[1])
-    gate_generator = of.FermionOperator(fop_string, 1.0)
-    gate_generator = 1j * (gate_generator - of.hermitian_conjugated(gate_generator))
-
-    return gate_generator
-
-
-def get_givens_gate(qubits: Tuple[cirq.Qid, ...], param: float) -> cirq.Operation:
-    """Get a the givens rotation gate on two qubits.
-
-    Args:
-        qubits: The two qubits to apply the gate to.
-        param: The parameter for the givens rotation.
-
-    Returns:
-        The givens rotation gate.
-    """
-    return cirq.givens(param).on(qubits[0], qubits[1])
-
-
-def get_layer_indices(
-    layer_spec: trial_wf.LayerSpec, n_elec: int
-) -> List[Tuple[int, int]]:
-    """Get the indices for the heuristic layers.
-
-    Args:
-        layer_spec: The layer specification.
-        n_elec: The number of electrons.
-
-    Returns:
-        A list of indices for the layer.
-    """
-    indices_generators = {
-        "in_pair": get_indices_heuristic_layer_in_pair(n_elec),
-        "cross_pair": get_indices_heuristic_layer_cross_pair(n_elec),
-        "cross_spin": get_indices_heuristic_layer_cross_spin(n_elec),
-    }
-    indices_generator = indices_generators[layer_spec.layout]
-
-    return [indices for indices in indices_generator]
-
-
-def get_layer_gates(
-    layer_spec: trial_wf.LayerSpec,
-    n_elec: int,
-    params: np.ndarray,
-    fermion_index_to_qubit_map: Dict[int, cirq.GridQubit],
-) -> List[cirq.Operation]:
-    """Gets the gates for a hardware efficient layer of the ansatz.
-
-    Args:
-        layer_spec: The layer specification.
-        n_elec: The number of electrons.
-        params: The variational parameters for the hardware efficient gate layer.
-        fermion_index_to_qubit_map: A mapping between fermion mode indices and qubits.
-
-    Returns:
-        A list of gates for the layer.
-    """
-
-    indices_list = get_layer_indices(layer_spec, n_elec)
-
-    gate_funcs = {"givens": get_givens_gate, "charge_charge": get_charge_charge_gate}
-    gate_func = gate_funcs[layer_spec.base_gate]
-
-    gates = []
-    for indices, param in zip(indices_list, params):
-        qubits = tuple(fermion_index_to_qubit_map[ind] for ind in indices)
-        gates.append(gate_func(qubits, param))
-
-    return gates
-
-
-def get_layer_generators(
-    layer_spec: trial_wf.LayerSpec, n_elec: int
-) -> List[of.FermionOperator]:
-    """Gets the generators for rotations in a hardware efficient layer of the ansatz.
-
-    Args:
-        layer_spec: The layer specification.
-        n_elec: The number of electrons.
-
-    Returns:
-        A list of generators for the layers.
-    """
-
-    indices_list = get_layer_indices(layer_spec, n_elec)
-
-    gate_funcs = {
-        "givens": get_givens_generator,
-        "charge_charge": get_charge_charge_generator,
-    }
-    gate_func = gate_funcs[layer_spec.base_gate]
-
-    return [gate_func(indices) for indices in indices_list]
-
-
-def get_heuristic_gate_generators(
-    n_elec: int, layer_specs: Sequence[trial_wf.LayerSpec]
-) -> List[of.FermionOperator]:
-    """Get gate generators for the heuristic ansatz.
-
-    Args:
-        n_elec: The number of electrons.
-        layer_specs: The layer specifications.
-
-    Returns:
-        A list of generators for the layers.
-    """
-    gate_generators = []
-
-    for layer_spec in layer_specs:
-        gate_generators += get_layer_generators(layer_spec, n_elec)
-
-    return gate_generators
-
-
-def get_heuristic_circuit(
-    layer_specs: Sequence[trial_wf.LayerSpec],
-    n_elec: int,
-    params: np.ndarray,
-    fermion_index_to_qubit_map: Dict[int, cirq.GridQubit],
-) -> cirq.Circuit:
-    """Get a circuit for the heuristic ansatz.
-
-    Args:
-        layer_specs: The layer specs for the heuristic layers.
-        n_elec: The number of electrons.
-        params: The variational parameters for the circuit.
-        fermion_index_to_qubit_map: A mapping between fermion mode indices and qubits.
-
-    Returns:
-        A circuit for the heuristic ansatz.
-    """
-    gates: List[cirq.Operation] = []
-
-    for layer_spec in layer_specs:
-        params_slice = params[len(gates) :]
-        gates += get_layer_gates(
-            layer_spec, n_elec, params_slice, fermion_index_to_qubit_map
-        )
-
-    return cirq.Circuit(gates)
 
 
 def orbital_rotation_gradient_matrix(
@@ -1167,7 +480,7 @@ def get_pp_plus_params(
     restricted: bool = False,
     random_parameter_scale: float = 1.0,
     initial_orbital_rotation: Optional[np.ndarray] = None,
-    heuristic_layers: Tuple[trial_wf.LayerSpec, ...],
+    heuristic_layers: Tuple[layer_spec.LayerSpec, ...],
     do_pp: bool = True,
     n_optimization_restarts: int = 1,
     do_print: bool = True,
@@ -1214,7 +527,7 @@ def get_pp_plus_params(
         err_msg = "use_fast_gradients does not work with initial orbital rotation."
         assert initial_orbital_rotation is None, err_msg
 
-    gate_generators = _get_pp_plus_gate_generators(
+    gate_generators = afqmc_generators.get_pp_plus_gate_generators(
         n_elec=n_elec, heuristic_layers=heuristic_layers, do_pp=do_pp
     )
